# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing courses and related resources",
        "title": "Courses Service API",
        "contact": {
            "name": "El mejor grupo de todos ndea deau",
            "url": "https://github.com/classconnect-grupo3",
            "email": "classconnectingsoft2@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/courses/{id}/enroll": {
            "post": {
                "description": "Enroll a student in a course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Enroll a student in a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enrollment request",
                        "name": "enrollmentRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.EnrollStudentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/courses/{id}/enrollments": {
            "get": {
                "description": "Get enrollments by course ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Get enrollments by course ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Enrollment"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{id}/favourite": {
            "post": {
                "description": "Set a course as favourite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Set a course as favourite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Favourite course request",
                        "name": "favouriteCourseRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.SetFavouriteCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SetFavouriteCourseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Unset a course as favourite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Unset a course as favourite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Unset favourite course request",
                        "name": "unsetFavouriteCourseRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UnsetFavouriteCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UnsetFavouriteCourseResponse"
                        }
                    }
                }
            }
        },
...
        "/courses/{id}/unenroll": {
            "delete": {
                "description": "Unenroll a student from a course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Unenroll a student from a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Unenrollment request",
                        "name": "unenrollmentRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UnenrollStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UnenrollStudentResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Enrollment": {
            "type": "object",
            "properties": {
                "completed_date": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "enrolled_at": {
                    "type": "string"
                },
                "favourite": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.EnrollmentStatus"
                },
                "student_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.EnrollmentStatus": {
            "type": "string",
            "enum": [
                "active",
                "dropped",
                "completed"
            ],
            "x-enum-varnames": [
                "EnrollmentStatusActive",
                "EnrollmentStatusDropped",
                "EnrollmentStatusCompleted"
            ]
        },
...
        "schemas.EnrollStudentRequest": {
            "type": "object",
            "required": [
                "student_id"
            ],
            "properties": {
                "student_id": {
                    "type": "string"
                }
            }
        },
...
        "schemas.SetFavouriteCourseRequest": {
            "type": "object",
            "required": [
                "student_id"
            ],
            "properties": {
                "student_id": {
                    "type": "string"
                }
            }
        },
        "schemas.SetFavouriteCourseResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "schemas.UnenrollStudentRequest": {
            "type": "object",
            "required": [
                "student_id"
            ],
            "properties": {
                "student_id": {
                    "type": "string"
                }
            }
        },
        "schemas.UnenrollStudentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "schemas.UnsetFavouriteCourseRequest": {
            "type": "object",
            "required": [
                "student_id"
            ],
            "properties": {
                "student_id": {
                    "type": "string"
                }
            }
        },
        "schemas.UnsetFavouriteCourseResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}